backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
boto3==1.38.17
botocore==1.38.17
celery==5.3.4
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==4.5.5
referencing==0.36.2
rpds-py==0.25.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
